backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
nltk
textblob
openai
python-dotenv
